QML: Port QV4::CompiledData::Binding to new special integer bitfield
Task-number: QTBUG-99545
(cherry picked from commit 36ebee4e69182f0e44d87691d4740b271e1dcf38)

+++ tests/auto/qml/qmldiskcache/tst_qmldiskcache.cpp
@@ -346,7 +346,7 @@ void tst_qmldiskcache::regenerateAfterChange()
 
         const QV4::CompiledData::Object *obj = qmlUnit->objectAt(0);
         QCOMPARE(quint32(obj->nBindings), quint32(1));
-        QCOMPARE(quint32(obj->bindingTable()->type), quint32(QV4::CompiledData::Binding::Type_Script));
+        QCOMPARE(obj->bindingTable()->type(), QV4::CompiledData::Binding::Type_Script);
         QCOMPARE(quint32(obj->bindingTable()->value.compiledScriptIndex), quint32(0));
 
         QCOMPARE(quint32(testUnit->functionTableSize), quint32(1));
@@ -374,7 +374,7 @@ void tst_qmldiskcache::regenerateAfterChange()
 
         const QV4::CompiledData::Object *obj = qmlUnit->objectAt(0);
         QCOMPARE(quint32(obj->nBindings), quint32(2));
-        QCOMPARE(quint32(obj->bindingTable()->type), quint32(QV4::CompiledData::Binding::Type_Number));
+        QCOMPARE(obj->bindingTable()->type(), QV4::CompiledData::Binding::Type_Number);
 
         const QV4::Value value(testUnit->constants()[obj->bindingTable()->value.constantValueIndex]);
         QCOMPARE(value.doubleValue(), double(42));
@@ -418,7 +418,7 @@ void tst_qmldiskcache::registerImportForImplicitComponent()
 
         const QV4::CompiledData::Object *obj = qmlUnit->objectAt(0);
         QCOMPARE(quint32(obj->nBindings), quint32(1));
-        QCOMPARE(quint32(obj->bindingTable()->type), quint32(QV4::CompiledData::Binding::Type_Object));
+        QCOMPARE(obj->bindingTable()->type(), QV4::CompiledData::Binding::Type_Object);
 
         const QV4::CompiledData::Object *implicitComponent = qmlUnit->objectAt(obj->bindingTable()->value.objectIndex);
         QCOMPARE(testUnit->stringAtInternal(implicitComponent->inheritedTypeNameIndex), QStringLiteral("QmlInternals.") + componentType.elementName());
